# src/main/java/com/example/bankcards/dto/api/auth/SignInResponse.java
package com.example.bankcards.dto.api.auth;

import io.swagger.v3.oas.annotations.media.Schema;
import lombok.Builder;
import lombok.Getter;

import java.time.Duration;

@Getter
@Builder
public class SignInResponse {

    @Schema(description = "JWT access token", example = "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...")
    private final String token;

    @Schema(description = "Refresh token for getting new access tokens", example = "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...")
    private final String refreshToken;

    @Schema(description = "Duration until access token expires, ISO-8601 duration format", example = "PT1H")
    private final Duration expirationTime;
}